bynx-backend/requirements.txt
fastapi
uvicorn
motor
pymysql
sqlalchemy
mysql-connector-python
pydantic[email]
passlib
python-dotenv